+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### [2.4.1] - 2026-05-13
 
-
-- after trapping a denysoft, return denysoft later (vs deny)
+- fix: after trapping a denysoft, return denysoft later (vs deny)
+- update spammy TLDs
+- fix: spamassassin hits -> score
 
 ### [2.4.0] - 2026-05-06